restructuring a bit
