+++ acquire_data/images/common_utils.py
