+++ acquire_data/images/convert_pkl_to_png.py
